requirements.txt
modal>=0.64.0
requests>=2.31.0
Pillow>=10.0.0